+++ modules/AWSPowerShell/Cmdlets/ConnectParticipant/Basic/Get-CONNPTranscript-Cmdlet.cs
@@ -28,11 +28,10 @@ using Amazon.ConnectParticipant.Model;
 namespace Amazon.PowerShell.Cmdlets.CONNP
 {
     /// <summary>
-    /// Retrieves a transcript of the session, including details about any attachments. Note
-    /// that ConnectionToken is used for invoking this API instead of ParticipantToken.
+    /// Retrieves a transcript of the session, including details about any attachments. 
     /// 
-    ///  
-    /// <para>
+    ///  <note><para><code>ConnectionToken</code> is used for invoking this API instead of <code>ParticipantToken</code>.
+    /// </para></note><para>
     /// The Amazon Connect Participant Service APIs do not use <a href="https://docs.aws.amazon.com/general/latest/gr/signature-version-4.html">Signature
     /// Version 4 authentication</a>.
     /// </para><br/><br/>This cmdlet automatically pages all available results to the pipeline - parameters related to iteration are only needed if you want to manually control the paginated output. To disable autopagination, use -NoAutoIteration.