+++ modules/AWSPowerShell/Cmdlets/ConnectParticipant/Basic/Send-CONNPMessage-Cmdlet.cs
@@ -28,11 +28,10 @@ using Amazon.ConnectParticipant.Model;
 namespace Amazon.PowerShell.Cmdlets.CONNP
 {
     /// <summary>
-    /// Sends a message. Note that ConnectionToken is used for invoking this API instead of
-    /// ParticipantToken.
+    /// Sends a message.
     /// 
-    ///  
-    /// <para>
+    ///  <note><para><code>ConnectionToken</code> is used for invoking this API instead of <code>ParticipantToken</code>.
+    /// </para></note><para>
     /// The Amazon Connect Participant Service APIs do not use <a href="https://docs.aws.amazon.com/general/latest/gr/signature-version-4.html">Signature
     /// Version 4 authentication</a>.
     /// </para>
@@ -66,7 +65,9 @@ namespace Amazon.PowerShell.Cmdlets.CONNP
         #region Parameter Content
         /// <summary>
         /// <para>
-        /// <para>The content of the message.</para>
+        /// <para>The content of the message. </para><ul><li><para>For <code>text/plain</code> and <code>text/markdown</code>, the Length Constraints
+        /// are Minimum of 1, Maximum of 1024. </para></li><li><para>For <code>application/json</code>, the Length Constraints are Minimum of 1, Maximum
+        /// of 12000. </para></li></ul>
         /// </para>
         /// </summary>
         #if !MODULAR
@@ -83,7 +84,8 @@ namespace Amazon.PowerShell.Cmdlets.CONNP
         #region Parameter ContentType
         /// <summary>
         /// <para>
-        /// <para>The type of the content. Supported types are text/plain.</para>
+        /// <para>The type of the content. Supported types are <code>text/plain</code>, <code>text/markdown</code>,
+        /// and <code>application/json</code>.</para>
         /// </para>
         /// </summary>
         #if !MODULAR
@@ -101,7 +103,9 @@ namespace Amazon.PowerShell.Cmdlets.CONNP
         /// <summary>
         /// <para>
         /// <para>A unique, case-sensitive identifier that you provide to ensure the idempotency of
-        /// the request.</para>
+        /// the request. If not provided, the Amazon Web Services SDK populates this field. For
+        /// more information about idempotency, see <a href="https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/">Making
+        /// retries safe with idempotent APIs</a>.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]